mch2022: More i2c fixes / workaround for readfrom_mem
In the previous version we just refused all transfers that don't
properly generate a STOP at the end. But the way the micropython
code is written the readfrom_mem command will generate a write without
stop then a read.

We can't support that as two different commands, we we detect that
access pattern, store the small write temporarely (it's the register
address), then merge it with the following read command when it
comes.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>