vfs_fat: Make the mess of path work for mch2022
So, we need to test for VFS_NATIVE_EXTERNAL_MP _first_ because
VFS_NATIVE_INTERNAL_MP is "/" so it will match all the time ...

Also because of that, for the internal case we need '-1' offset when
building the final path but _not_ for the external case.

It's not a great fix (because changing VFS_NATIVE_INTERNAL_MP to
anything else than '/' will break) but it's good enough for it to
work for MCH2022.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>