chore/micropython: Update MicroPython (v1.20.0~33b403df)
This updates MicroPython to v1.20.0+git33b403df.  This update brings
better support for inclusion of MicroPython as an ESP-IDF component.

At the most basic level, one can now define a stub "component" as
follows to build MicroPython:

    # Component-level cmake file for building MicroPython on ESP32.
    cmake_minimum_required(VERSION 3.12)

    if(NOT DEFINED MICROPY_DIR)
        get_filename_component(MICROPY_DIR ${PROJECT_DIR}/external/micropython ABSOLUTE)
    endif()

    # Include core source components.
    include(${MICROPY_DIR}/ports/esp32/main/CMakeLists.txt)

After setting the correct path to `MICROPY_DIR` and including the
***upstream*** component's `CMakeLists.txt` file, CMake is able to
correctly identify the ESP-IDF compatible component and perform the
build.