build/Makefile: Remove hard-coded boards
This removes many of the hard coded board definitions in favor of
defining a `BOARD` environment variable.  This should correspond to a
directory located at one of the following two paths:

- `boards/`
- `external/micropython/ports/esp32/boards/`

In an attempt to make understanding the path more ideomatic we present a
symbolic link to the second entry in this list at:

- `boards/upstream/`

Additionally we define a board, `DEFAULT`, which builds the
`badgePython` firmware for a basic ESP32 target with an SSD1306 OLED
screen.  This target should allow for basic [Wokwi][1] testing from CI
builds.

Finally a make target was added for `mpy_cross` which allows `make`
users to correctly detect if the dependency on `mpy-cross` is correctly
met.

[1]: https://wokwi.com/