external/micropython: Relocate and document sub-module
When acting as a code historian across this repository it was very
challenging to discover the provenance of many of the files.  Often they
(seem) to have been copy-pasta which then lost the history and/or
context of why certain changes had been made.

This isolates upstream MicroPython from any source or configuration
changes we may need to make.  The architecture is that we consume
MicroPython as an [ESP-IDF "component"][1].  This then allows us to use
the ESP-IDF to mutate the build as needed.  Under the hood the ESP-IDF
is implementing these "[APIs][2]" (as they call them) as functions
within CMake.

Thus we now have an ESP-IDF "component" named `micropython`.  The
definition of the component via the API is currently located at:

[`external/micropython/ports/esp32/main/CMakeLists.txt#162-L180`][3]:

```cmake
idf_component_register(
    SRCS
        ${MICROPY_SOURCE_PY}
        ${MICROPY_SOURCE_EXTMOD}
        ${MICROPY_SOURCE_SHARED}
        ${MICROPY_SOURCE_LIB}
        ${MICROPY_SOURCE_DRIVERS}
        ${MICROPY_SOURCE_PORT}
        ${MICROPY_SOURCE_BOARD}
    INCLUDE_DIRS
        ${MICROPY_INC_CORE}
        ${MICROPY_INC_USERMOD}
        ${MICROPY_PORT_DIR}
        ${MICROPY_BOARD_DIR}
        ${CMAKE_BINARY_DIR}
    REQUIRES
        ${IDF_COMPONENTS}
)
```

As a result it is solely code maintained by this project which comprises
the definition of `micropython` as we consume it in
`components/micropython`.

[1]: https://docs.espressif.com/projects/esp-idf/en/v4.4.5/esp32/api-guides/build-system.html#using-third-party-cmake-projects-with-components
[2]: https://docs.espressif.com/projects/esp-idf/en/v4.4.5/esp32/api-guides/build-system.html#esp-idf-cmake-build-system-api
[3]: https://github.com/micropython/micropython/blob/master/ports/esp32/main/CMakeLists.txt#L162-L180