typo/gitignore: Missing root scope on files
There were some files which had been added to gitignore, but were not
under the root scope.

The following files in gitignore should only ignored at the
root of the repository.  Their existence at other paths within the
repository is not only acceptable, but needed for our new `boards`
subdirectory:

- sdkconfig
- partitions.csv
- sdkconfig.old