improvement/libopus: Vendor upstream module
There were a number of issues which had been solved in upstream commits.
There are a bunch of buffer overflows which have been fixed, various
issues resolved during valgrind testing, etc.  All of this should lead
to fewer issues in the long run with memory leaks.

This commit should atomically remove the previous vendored source code,
introduce the upstream repository as a git submodule, and update the
CMakeLists.txt file to import the upstream definition.  This is a good
example for users who may want to add their own third party libraries

From here on out (with the exception of needing to carry a patch set for
some reason which gets applied locally), we can now update the libopus
component using a normal git submodule update process.  Since our
CMakeLists.txt in `components/libopus` exists primarily to
configure the desired options as to how the build is performed this
separates the concerns of the project: our "opinions" on the build
exist in `components`.  The interface for where we stop maintaining code
is at `external`.