mpy/QSTR: Make nvs & rtcmem USER_C_MODULES
This affects the following components:
- RTC memory driver
- NVS driver
- Micropython module for NVS
- Micropython module for RTC memory
- Micropython module for external modules

It looks like previously most of the badgePython C code for interfacing
with Python were implemented as "extmod"s in MicroPython.  From a
maintenance perspective MicroPython intends an "extmod" to live within
the MicroPython codebase, but be something which affects multiple
ports... for example the use of mbedTLS or TinyUSB.

To clarify the relationship between badgePython and MicroPython, lets
review part of the MicroPython documentation: [Implementing a
Module][1]:

> This chapter details how to implement a core module in MicroPython.
> MicroPython modules can be one of the following:
>
> -   Built-in module: A general module that is be part of the MicroPython
>     repository.
> -   User module: A module that is useful for your specific project that
>     you maintain in your own repository or private codebase.
> -   Dynamic module: A module that can be deployed and imported at
>     runtime to your device.
>
> A module in MicroPython can be implemented in one of the following
> locations:
>
> -   py/: A core library that mirrors core CPython functionality.
> -   extmod/: A CPython or MicroPython-specific module that is shared
>     across multiple ports.
> -   ports/<port>/: A port-specific module.

Clearly a "User module" is the most appropriate term for what we are
trying to implment.  This commit migrates `nvs` and `driver_rtcmem` to
be "User modules" and moves them into the `components` directory.

To understand what this sort of change entails, let's review the
MicroPython documentation for "Extending MicroPython in C", specifically
"[Compiling the cmodule into MicroPython][2]".

In my opinion the MicroPython documentation fails to drive a very
important point home:

`USER_C_MODULES` mechanism implemented in MicroPython expects to receive
a CMake script which then uses `include()` to enumerate all of the project
modules as CMake scripts (as a specific path) or CMake modules (found via
searching for the module in the locations specified in the
`CMAKE_MODULE_PATH` CMake variable).  This is why this commit adds the
file `components/external_micropy_modules.cmake`.  It allows the end
user to specify the modules that the specifically wish to import into
the MicroPython firmware.

In the long run, this will be improved to support the `BOARDS` mechanism
allowing for easy driver selection, integrated in with Kconfig.

Next, for each of our drivers we add a CMake script which will allow
MicroPython to execute the QSTR macro generation process.  If you find
yourself troubleshooting missing definitions of `MP_QSTR_` macros, this
is the file to look at.  I'm using a convention of calling the file
`micropy_module.cmake` to distinguish it from the CMake script which
will be used to build the module.

The `micropy_module.cmake` file is a CMake script which defines the
following:

- The components source code as a target of type `INTERFACE`
- The components source directory as an include location of type
  `INTERFACE`
- That this is component should be attached to the `usermod` target as
  an `INTERFACE`.

This is pretty much a rote implementation of the example provided in
the [documentation][2] under "Structure of an external C module".
Towards the end of this section CMake based implementations are
discussed.

[1]: https://docs.micropython.org/en/v1.20.0/develop/library.html
[2]: https://docs.micropython.org/en/v1.20.0/develop/cmodules.html#compiling-the-cmodule-into-micropython